+++ src/features/MergeNetworks/utils/attributesOperations.tsx
@@ -41,3 +41,14 @@ export function addMergedAtt(castedRecord: Record<string, ValueType>, oriRecord:
     return castedRecord;
 
 }
+
+export function attributeValueMatcher(val: ValueType, nodeAttMap: Map<IdType, ValueType>): string {
+    if (val !== undefined) {
+        for (const entry of nodeAttMap.entries()) {
+            if (entry[1] !== '' && val === entry[1]) {
+                return entry[0];
+            }
+        }
+    }
+    return ''
+}
