--- src/features/MergeNetworks/utils/valueMatcher.tsx
@@ -1,13 +1,0 @@
-import { IdType } from "../../../models/IdType";
-import { Table, ValueType } from "../../../models/TableModel";
-
-export function attributeValueMatcher(val: ValueType, nodeAttMap: Map<IdType, ValueType>): string {
-    if (val !== undefined) {
-        for (const entry of nodeAttMap.entries()) {
-            if (entry[1] !== '' && val === entry[1]) {
-                return entry[0];
-            }
-        }
-    }
-    return ''
-}
